Using template-based pattern for OrderedParquetDataset to derive ReadOnlyOrderedParquetDataset.
